stock-news-bot/requirements-reel.txt
Pillow
requests
beautifulsoup4
moviepy>=2.0
yfinance
matplotlib
numpy
imageio[ffmpeg]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
